Refactor streaming related classes to a dedicated streaming directory
